Add spend request parsing

+++ budget/src/pay_tax_request.cpp
@@ -4,8 +4,7 @@ using namespace std;
 
 void PayTaxRequest::ParseFrom(string_view input) {
   date_from = Date::FromString(ReadToken(input));
-  //  date_to = Date::FromString(input);
   date_to = Date::FromString(ReadToken(input));
-  percentage = ConvertToInt(input);
+  percentage = static_cast<size_t>(ConvertToInt(input));
 }
